distributed
改成了分布式，使用请安装redis和celery,注意设置延时
